project_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
pyjwt==2.9.0
pydantic==2.9.2
pydantic[email]==2.9.2
bcrypt==4.2.0
python-multipart==0.0.9
alembic==1.13.3
requests==2.30.0